Changed compiler options to report most warnings. Fixed a number of minor issues related to warnings.

+++ src/l1j/server/server/datatables/SpawnTable.java
@@ -195,7 +195,7 @@ public class SpawnTable {
 		if (rate == 1 || npc.isAmountFixed()) {
 			return count;
 		} else {
-			return NumberUtil.randomRound((double) (count * rate));
+			return NumberUtil.randomRound(count * rate);
 		}
 	}
 }
